--- a/osu.Game/Rulesets/Objects/HitObjectParser.cs
+++ b/osu.Game/Rulesets/Objects/HitObjectParser.cs
@@ -1,10 +1,12 @@
 ﻿// Copyright (c) ppy Pty Ltd <contact@ppy.sh>. Licensed under the MIT Licence.
 // See the LICENCE file in the repository root for full licence text.
 
+using System;
+
 namespace osu.Game.Rulesets.Objects
 {
     public abstract class HitObjectParser
     {
-        public abstract HitObject? Parse(string text);
+        public abstract HitObject? Parse(ReadOnlySpan<char> text);
     }
 }