… CodeAnalysis/BannedSymbols.txt