Supersedes #23282 .

This PR now uses `MemoryExtensions.Split(ReadOnlySpan<char>)` API in .NET 8, which splits a span of chars into pre-determined sections with no object allocation.