@smoogipoo 2024-07-09
Rough benchmark:

```
Master:

| Method                  | Mean     | Error     | StdDev    | Gen0     | Gen1     | Allocated |
|------------------------ |---------:|----------:|----------:|---------:|---------:|----------:|
| BenchmarkBundledBeatmap | 2.676 ms | 0.0315 ms | 0.0263 ms | 691.4063 | 343.7500 |   4.16 MB |

PR:

| Method                  | Mean     | Error     | StdDev    | Gen0     | Gen1     | Allocated |
|------------------------ |---------:|----------:|----------:|---------:|---------:|----------:|
| BenchmarkBundledBeatmap | 2.793 ms | 0.0326 ms | 0.0273 ms | 585.9375 | 292.9688 |   3.52 MB |
```