@huoyaoyuan 2024-07-09
Hmm, the lower GC pressure doesn't result in better net performance like the previous PR. I'll do some profiling later.