@smoogipoo 2024-07-09
It's going to be bounded by all the other much larger allocs, `HitObject`s, etc... I found the same results re: performance with my original attempt at this.