@huoyaoyuan 2024-07-09
> See decoding test failures (https://github.com/ppy/osu/pull/28779/checks?check_run_id=27174798514) if you haven't already :)

I'm mainly working for OSS code in evening. I have my day-time work now :)