@huoyaoyuan 2024-07-09
On my machine:

main:
| Method                  | Mean     | Error     | StdDev    | Gen0     | Gen1     | Allocated |
|------------------------ |---------:|----------:|----------:|---------:|---------:|----------:|
| BenchmarkBundledBeatmap | 1.581 ms | 0.0307 ms | 0.0315 ms | 230.4688 | 193.3594 |   4.15 MB |

PR:
| Method                  | Mean     | Error     | StdDev    | Gen0     | Gen1     | Allocated |
|------------------------ |---------:|----------:|----------:|---------:|---------:|----------:|
| BenchmarkBundledBeatmap | 1.493 ms | 0.0149 ms | 0.0132 ms | 193.3594 | 167.9688 |    3.5 MB |

It may be some environment or GC difference across platforms.